License Keys and Plugin Recommendations and necessary auxiliary changes
Because of the need to deactivate license keys when the plugin is
deactivated, we were forced to add Zycache settings to the plugin so
that a user doesn't need to deactivate the plugin to turn off zycache
(as that would deactivate licenses).

Because deactivating the plugin means that you're no longer a Zysys
Client, we edited the deactivation procedure to undo all of the
optimizations we use in the plugin.

Additionally, I combined the license-keys.php file and
plugin-updater.php into the main plugin to ensure that it can't be
accidentally overwritten.