Created a hook for maintenance to better integrate with the modular idea
Hook: 'zysyshosting_maintenance_action' allows for the modular system to
keep function calls inside the file they were defined in so that we
can see that the new security measure (for example) is actually being
installed during maintenance.

This should make the modular system much easier to maintain.  Now we
just have to organize everything!